Create output module, some renaming
