+++ config/osc_config.yml
@@ -7,8 +7,8 @@ displays:
 ---
 sound:
   neuron_notes:
-    ip: 127.0.0.1
-    port: 1337
+    ip: 192.168.33.1
+    port: 8080
     address: /neuron_notes/spikes/
 
 ---
@@ -16,4 +16,4 @@ sources:
   simulator:
     ip: 192.168.33.1
     port: 8080
-    address: '/spikes'
+    address: /spikes
